// src/api/back/models/BracketMatchDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Public REST api for dota2classic
 * All stuff
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    MatchGameDto,
    MatchGameDtoFromJSON,
    MatchGameDtoFromJSONTyped,
    MatchGameDtoToJSON,
    MatchStatus,
    MatchStatusFromJSON,
    MatchStatusFromJSONTyped,
    MatchStatusToJSON,
    ParticipantResultDto,
    ParticipantResultDtoFromJSON,
    ParticipantResultDtoFromJSONTyped,
    ParticipantResultDtoToJSON,
} from './';

/**
 * 
 * @export
 * @interface BracketMatchDto
 */
export interface BracketMatchDto {
    /**
     * 
     * @type {MatchStatus}
     * @memberof BracketMatchDto
     */
    status: MatchStatus;
    /**
     * 
     * @type {number}
     * @memberof BracketMatchDto
     */
    id: number;
    /**
     * 
     * @type {number}
     * @memberof BracketMatchDto
     */
    stageId: number;
    /**
     * 
     * @type {number}
     * @memberof BracketMatchDto
     */
    groupId: number;
    /**
     * 
     * @type {number}
     * @memberof BracketMatchDto
     */
    roundId: number;
    /**
     * 
     * @type {number}
     * @memberof BracketMatchDto
     */
    childCount: number;
    /**
     * 
     * @type {number}
     * @memberof BracketMatchDto
     */
    number: number;
    /**
     * 
     * @type {ParticipantResultDto}
     * @memberof BracketMatchDto
     */
    opponent1?: ParticipantResultDto;
    /**
     * 
     * @type {ParticipantResultDto}
     * @memberof BracketMatchDto
     */
    opponent2?: ParticipantResultDto;
    /**
     * 
     * @type {string}
     * @memberof BracketMatchDto
     */
    startDate: string;
    /**
     * 
     * @type {Array<MatchGameDto>}
     * @memberof BracketMatchDto
     */
    games: Array<MatchGameDto>;
}

export function BracketMatchDtoFromJSON(json: any): BracketMatchDto {
    return BracketMatchDtoFromJSONTyped(json, false);
}

export function BracketMatchDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): BracketMatchDto {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'status': MatchStatusFromJSON(json['status']),
        'id': json['id'],
        'stageId': json['stage_id'],
        'groupId': json['group_id'],
        'roundId': json['round_id'],
        'childCount': json['child_count'],
        'number': json['number'],
        'opponent1': !exists(json, 'opponent1') ? undefined : ParticipantResultDtoFromJSON(json['opponent1']),
        'opponent2': !exists(json, 'opponent2') ? undefined : ParticipantResultDtoFromJSON(json['opponent2']),
        'startDate': json['startDate'],
        'games': ((json['games'] as Array<any>).map(MatchGameDtoFromJSON)),
    };
}

export function BracketMatchDtoToJSON(value?: BracketMatchDto | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'status': MatchStatusToJSON(value.status),
        'id': value.id,
        'stage_id': value.stageId,
        'group_id': value.groupId,
        'round_id': value.roundId,
        'child_count': value.childCount,
        'number': value.number,
        'opponent1': ParticipantResultDtoToJSON(value.opponent1),
        'opponent2': ParticipantResultDtoToJSON(value.opponent2),
        'startDate': value.startDate,
        'games': ((value.games as Array<any>).map(MatchGameDtoToJSON)),
    };
}