// src/api/back/models/OpponentResult.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Public REST api for dota2classic
 * All stuff
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * 
 * @export
 * @enum {string}
 */
export enum OpponentResult {
    Win = 'win',
    Loss = 'loss',
    Draw = 'draw'
}

export function OpponentResultFromJSON(json: any): OpponentResult {
    return OpponentResultFromJSONTyped(json, false);
}

export function OpponentResultFromJSONTyped(json: any, ignoreDiscriminator: boolean): OpponentResult {
    return json as OpponentResult;
}

export function OpponentResultToJSON(value?: OpponentResult | null): any {
    return value as any;
}